+++ infra/terraform/azure/main.tf
@@ -1,0 +1,41 @@
+resource "azurerm_resource_group" "this" {
+  name     = "${var.name}-rg"
+  location = var.location
+}
+
+resource "azurerm_virtual_network" "this" {
+  name                = "${var.name}-vnet"
+  location            = azurerm_resource_group.this.location
+  resource_group_name = azurerm_resource_group.this.name
+  address_space       = var.vnet_cidr
+}
+
+resource "azurerm_subnet" "aks" {
+  name                 = "${var.name}-aks"
+  resource_group_name  = azurerm_resource_group.this.name
+  virtual_network_name = azurerm_virtual_network.this.name
+  address_prefixes     = var.subnet_cidr
+}
+
+resource "azurerm_kubernetes_cluster" "this" {
+  name                = var.name
+  location            = azurerm_resource_group.this.location
+  resource_group_name = azurerm_resource_group.this.name
+  dns_prefix          = var.name
+
+  default_node_pool {
+    name           = "system"
+    node_count     = var.node_count
+    vm_size        = var.node_vm_size
+    vnet_subnet_id = azurerm_subnet.aks.id
+  }
+
+  identity {
+    type = "SystemAssigned"
+  }
+
+  network_profile {
+    network_plugin = "azure"
+    network_policy = "azure"
+  }
+}