+++ infra/terraform/azure/outputs.tf
@@ -1,0 +1,11 @@
+output "cluster_name" {
+  value = azurerm_kubernetes_cluster.this.name
+}
+
+output "resource_group" {
+  value = azurerm_resource_group.this.name
+}
+
+output "kubectl_config_command" {
+  value = "az aks get-credentials --resource-group ${azurerm_resource_group.this.name} --name ${azurerm_kubernetes_cluster.this.name}"
+}